+++ enph253/poopy_mayonnaise/src/TapeFollow2Disc.hpp
@@ -27,7 +27,7 @@ private:
     const int reset;
     Tinah &tinah;
     int activePins[4];
-    double readings[4];    // readings
+    int readings[4];    // readings
     int tape[3][4];        // tape[0] : current read, tape[1] : previous, ...
     int timeMap[3];
     double errorPrev;
